+++ src/dashboard-front/src/App.vue
@@ -92,7 +92,7 @@
             <div class="header-aside-wrap">
               <LanguageToggle />
               <ProductInfo />
-              <UserInfoV2 v-if="isBkUserNameConfigured" />
+              <UserInfo v-if="isBkUserNameConfigured" />
             </div>
           </div>
         </template>
@@ -107,7 +107,7 @@
 <script lang="ts" setup>
 import LanguageToggle from '@/components/language-toggle/Index.vue';
 import ProductInfo from '@/components/product-info/Index.vue';
-import UserInfoV2 from '@/components/user-info-v2/Index.vue';
+import UserInfo from '@/components/user-info/Index.vue';
 import LogoWithoutTitle from '@/images/APIgateway-logo.png';
 import En from '../node_modules/bkui-vue/dist/locale/en.esm.js';
 import ZhCn from '../node_modules/bkui-vue/dist/locale/zh-cn.esm.js';
