+++ src/dashboard-front/src/locales/cn.json
@@ -2517,5 +2517,6 @@
   "长度需要在 1 到 4096 个字符之间": "长度需要在 1 到 4096 个字符之间",
   "规则已存在": "规则已存在",
   "请求被拦截时返回的消息": "请求被拦截时返回的消息",
-  "开发者中心": "开发者中心"
+  "开发者中心": "开发者中心",
+  "个人设置": "个人设置"
 }
